Move more implementations to metastore module
