+++ presto-hive-metastore/src/main/java/com/facebook/presto/hive/LocationHandle.java
