+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveErrorCode.java
@@ -35,12 +35,12 @@ public enum HiveErrorCode
     HIVE_BAD_DATA(7, EXTERNAL),
     HIVE_PARTITION_SCHEMA_MISMATCH(8, EXTERNAL),
     HIVE_MISSING_DATA(9, EXTERNAL),
-    HIVE_INVALID_PARTITION_VALUE(10, EXTERNAL),
+    // HIVE_INVALID_PARTITION_VALUE(10, EXTERNAL) moved to MetastoreErrorCode
     HIVE_TIMEZONE_MISMATCH(11, EXTERNAL),
     // HIVE_INVALID_METADATA(12) moved to MetastoreErrorCode
     HIVE_INVALID_VIEW_DATA(13, EXTERNAL),
     HIVE_DATABASE_LOCATION_ERROR(14, EXTERNAL),
-    HIVE_PATH_ALREADY_EXISTS(15, EXTERNAL),
+    // HIVE_PATH_ALREADY_EXISTS(15, EXTERNAL) moved to MetastoreErrorCode
     // HIVE_FILESYSTEM_ERROR(16) moved to MetastoreErrorCode
     // code HIVE_WRITER_ERROR(17) is deprecated
     HIVE_SERDE_NOT_FOUND(18, EXTERNAL),
@@ -60,11 +60,11 @@ public enum HiveErrorCode
     HIVE_TABLE_READ_ONLY(32, USER_ERROR),
     HIVE_PARTITION_NOT_READABLE(33, USER_ERROR),
     HIVE_TABLE_NOT_READABLE(34, USER_ERROR),
-    HIVE_TABLE_DROPPED_DURING_QUERY(35, EXTERNAL),
+    // HIVE_TABLE_DROPPED_DURING_QUERY(35, EXTERNAL) moved to MetastoreErrorCode
     // HIVE_TOO_MANY_BUCKET_SORT_FILES(36) is deprecated
-    HIVE_CORRUPTED_COLUMN_STATISTICS(37, EXTERNAL),
+    // HIVE_CORRUPTED_COLUMN_STATISTICS(37, EXTERNAL) moved to MetastoreErrorCode
     HIVE_EXCEEDED_SPLIT_BUFFERING_LIMIT(38, USER_ERROR),
-    HIVE_UNKNOWN_COLUMN_STATISTIC_TYPE(39, INTERNAL_ERROR),
+    // HIVE_UNKNOWN_COLUMN_STATISTIC_TYPE(39, INTERNAL_ERROR) moved to MetastoreErrorCode
     HIVE_TABLE_BUCKETING_IS_IGNORED(40, USER_ERROR),
     HIVE_TRANSACTION_NOT_FOUND(41, INTERNAL_ERROR),
     /**/;
